requirements.txt
pillow==11.1.0
tqdm==4.67.1
loguru==0.7.3
python-dotenv==1.0.1
wandb==0.19.4
torch==2.5.1
torchvision==0.20.1
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
scikit_learn==1.6.0
matplotlib==3.10.0
hydra-core==1.3.2
pytest==8.3.4
coverage==7.6.10
invoke==2.2.0
transformers==4.48.0
omegaconf==2.3.0
dvc-gs==3.0.1
numpy==2.2.1
pylint==3.3.3
kaggle==1.6.17